Added support build settings v2
