+++ Plugins/StoryGraphPlugin/Source/StoryGraphPluginEditor/Private/DetailPanelCustomizationNodes.cpp
@@ -6,9 +6,7 @@
 #include "DetailWidgetRow.h"
 #include "DetailCategoryBuilder.h"
 #include "StoryGraphObject.h"
-#include "Runtime/Engine/Classes/EdGraph/EdGraph.h"
 #include "GraphSchema_StoryGraph.h"
-#include "StoryGraph.h"
 #include "CustomNodes.h"
 #include "DetailLayoutBuilder.h"
 #include "IDetailPropertyRow.h"
