+++ Plugins/StoryGraphPlugin/Source/StoryGraphPluginEditor/Private/GraphSchema_StoryGraph.cpp
@@ -4,16 +4,14 @@
 #include "AssetEditor_StoryGraph.h"
 #include "ConnectionDrawingPolicy_StoryGraph.h"
 #include "CustomNodes.h"
-#include "EdGraph/EdGraph.h"
-#include "EdGraph/EdGraphPin.h"
-#include "Framework/MultiBox/MultiBoxBuilder.h"
-#include "GenericCommands.h"
-#include "Graph_StoryGraph.h"
 #include "GraphEditorActions.h"
+#include "Graph_StoryGraph.h"
 #include "ProxyNodes.h"
-#include "ScopedTransaction.h"
 #include "StoryGraph.h"
 #include "StoryGraphObject.h"
+#include "EdGraph/EdGraph.h"
+#include "EdGraph/EdGraphPin.h"
+#include "Runtime/Slate/Public/Framework/Commands/GenericCommands.h"
 #include "ToolMenus/Public/ToolMenu.h"
