+++ Plugins/StoryGraphPlugin/Source/StoryGraphPluginEditor/Public/AssetFactory_StoryGraph.h
@@ -1,6 +1,5 @@
 // Copyright 2016 Dmitriy Pavlov
 #pragma once
-#include "Engine/DataAsset.h"
 #include "Factories/Factory.h"
 #include "AssetFactory_StoryGraph.generated.h"
