+++ Plugins/StoryGraphPlugin/Source/StoryGraphPluginEditor/Public/StoryGraphPluginEditor.h
@@ -3,15 +3,6 @@
 #pragma once
 
 #include "EngineMinimal.h"
-#include "ModuleManager.h"
-//#include "StoryGraphPluginRuntimeClasses.h"
-//#include "EditorStyleSet.h"
-//#include "EdGraph/EdGraphNodeUtils.h"
-//#include "ScopedTransaction.h"
-//#include "EdGraph/EdGraphPin.h"
-
-
-
 
 class FStoryGraphEditorModule : public IModuleInterface
 {
