+++ Source/StoryGraphEditor.Target.cs
@@ -7,6 +7,7 @@ public class StoryGraphEditorTarget : TargetRules
 {
 	public StoryGraphEditorTarget(TargetInfo Target) : base(Target)
     {
+	    DefaultBuildSettings = BuildSettingsVersion.V2;
 		Type = TargetType.Editor;
         ExtraModuleNames.AddRange(new string[] { "StoryGraph" });
     }
